(halfORM) Catch error if there is some wierd inheritance in PG that Python MRO can't handle.
Sometimes the Model.get_relation_class crashes with the error :
Cannot create a consistent method resolution order (MRO) for bases ...

It seems to be a wrong inheritance definition in PostgreSQL relation.
It occurs for example on table b with the following definition:
- create table base(...);
- create table a(...) inherits(base);
- create table b(...) inherits(base, a);

The error is now reported but halfORM continues its work on the other
relations.