[hop][pylint] Ignore invalid-name and attribute-defined-outside-init.
A relation class is frozen. No attribute can be set if it's not a column of the
relation in the database. Yet pylint is not aware of the attributes which have
been set by the class-factory.
As well, the attribute names don't have to comply to the Python naming convention
as they are set in the postgresql database.